+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # 0.0.2
 
+- Add Flow
+- Add ESLint
 - Add component ScrollView
+- Add View property `accessibilityLabel`